requirements.txt
contourpy==1.0.5
cycler==0.11.0
Faker==14.2.0
fonttools==4.37.2
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.6.0
med2image==2.6.6
nibabel==4.0.2
numpy==1.23.3
packaging==21.3
parso==0.8.3
pfdicom==1.7.10
pfdicom-tagExtract==2.2.32
pfmisc==2.2.10
pftree==2.2.0
Pillow==9.2.0
pudb==2022.1.2
pydicom==2.3.0
pydicom-ext==0.6.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
scipy==1.9.1
urwid==2.1.2
urwid-readline==0.13
